requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==21.2.0
argon2-cffi-bindings==21.2.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.15.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
comtypes==1.4.10
confection==0.1.5
cymem==2.0.11
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.3
exceptiongroup==1.2.2
fastapi==0.115.11
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.16.1
fsspec==2025.3.0
future==1.0.0
gradio==5.20.1
gradio_client==1.7.2
groovy==0.1.2
groq==0.18.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.4.2
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.2.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
mpmath==1.3.0
murmurhash==1.0.12
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==3.4.2
nltk==3.9.1
notebook==6.4.6
numpy==2.0.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.6
playsound==1.2.2
preshed==3.0.9
prometheus-client==0.12.0
prompt-toolkit==3.0.24
PyAudio==0.2.14
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.0.6
pypiwin32==223
pyrsistent==0.18.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.1
pywin32==302
pywinpty==1.1.6
PyYAML==6.0.2
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
rich==13.9.4
ruff==0.9.10
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.1
srsly==2.5.1
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
terminado==0.12.1
testpath==0.5.0
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
tornado==6.1
tqdm==4.67.1
traitlets==5.1.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.27.0
wasabi==1.1.3
wcwidth==0.2.5
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
widgetsnbextension==3.5.2
wrapt==1.17.2
